requirements.txt
# =============================================================================
# Biomedical Equipment Troubleshooting Agent - Dependencies
# =============================================================================

# Core Framework
langgraph>=0.4.10
langchain-core>=0.3.0
langchain-community>=0.3.0

# LLM Integration
langchain-groq>=0.2.0
langchain-openai>=0.2.0

# Embeddings (Local)
sentence-transformers>=2.2.0

# Vector Database
chromadb>=0.5.0

# USB Multimeter Support
pyserial>=3.5

# PDF Support (for knowledge ingestion)
pypdf>=4.0.0

# Observability
langsmith>=0.1.0

# Configuration
python-dotenv>=1.0.0
pydantic>=2.0.0
pyyaml>=6.0

# CLI
colorama>=0.4.6

# Development
pytest>=8.0.0
pytest-asyncio>=0.23.0
